Pass the LSN to encrypt_block() explicitly.
The function should not read the LSN (to be used as the encryption IV) from
the page being encrypted because the input data can be in a shared buffer
locked only in shared mode. That makes it possible for backends to change the
LSN (due to hint bit updates) while encrypt_block() is reading it, so the IV
used for encryption is less deterministic (and therefore more prone to
duplicate values).