Reworked encryption of temporary files.
So far we could only construct the initialization vector (IV) out of
information that we could retrieve during encryption (such as block number or
PID of the writing process), but that approach is unwieldy when it comes to
shared temporary files (e.g. parallel worker creates temp file that the leader
process needs to decrypt). Another problem is that, if block number is used to
construct the IV, we need to pay special attention when appending one temp
file to another because that way the blocks of the appended (i.e. non-first)
file(s) get new numbers. In order to be able to decrypt the new file, we need
to keep the original block numbers in memory. Finally, we had to use the same
IV for given block even if it got overwritten any number of times.

This patch adds a header to each BLCKSZ-sized block of the temporary file,
which contains the IV. Thus the reader simply reads the IV from the file, w/o
having to reconstruct the IV from the information that the writer (possibly
another backend) used during encryption. And of course it's pretty easy now to
generate a new IV for each encryption run of the same block, which increases
the security.

The addition of the page header requires mapping between the "physical" offset
in the temporary file segment (i.e. where the data is located on disk) and the
"logical" offset (i.e. where the user, who is not aware of the headers, thinks
the data should be located). This mapping is transparent, meaning that the
user of the API does not care about the "physical" offsets. Naturally,
functions like BufFileTell() or BufFileSeek() work with the "logical" offset.

Since this rework breaks compatibility for the statistics files, I've bumped
the TDE version to 1.1. For the same reason this change cannot be applied to
minor PG releases which already have the TDE merged.

I also removed the regression tests, they'll be maintained as a separate
extension. The problem here is that two different sets of tests are needed -
one for encrypted instance and one for unencrypted one. I could not figure out
how to pick the correct set of tests automatically.