Make sure that overwriting of a temp file does not erase valid data.
When a buffer of an encrypted temporary file has been dumped, the following
block (if it exists) must be loaded into the buffer before we start writing
into it. Otherwise the next dump can overwrite valid data of the following
block with zeroes.

Fortunately this bugs seems to be hidden by the fact that logtape.c only
writes whole blocks and that other callers of the buffile.c API do not seem to
overwrite the temporary files at all.