Added an option to control the encryption key length.
So far, the key length was hard-wired to 128 bits. Now, the user can use the
--key-bits option to tell the initdb application which key length he
prefers. The accepted values are 128 (the default), 192 and 256 bits.

The key length information is stored in the control file, but also in the KDF
parameters file (global/kdf_params) because the pg_keytool utility may be
needed by initdb when the control file does not exist yet. This format change
implies that the KDF file needs to be migrated by pg_upgrade.