Reset the statistics if the statistics file is corrupt.
The particular problem was an encrypted statistics file whose size was not a
whole multiple of BLCKSZ (typically 8 kB). Although there can be other kinds
of corruption, this one is very easy to identify and handle. (It actually had
to be fixed for the regression tests to pass.)

The "transient BufFile" API had to be adjusted so that the caller can
recognize the problem by checking the function return value rather than by
receiving ERROR. The problem is that elog setup is not complete at the time
the statistics are read, so ERROR gets propagated to FATAL (causing the stats
collector to exit) and thus cannot be caught using the PG_TRY construct.