Reset the statistics if the statistics file is corrupt.
The particular problem was an encrypted statistics file whose size didn't meet
the restriction of the file format. Although there can be other kinds of
corruption, this one is very easy to identify and handle.

The "transient BufFile" API had to be adjusted so that the caller can
recognize the problem by checking the function return value rather than by
receiving ERROR. The problem is that elog setup is not complete at the time
the statistics are read, so ERROR gets propagated to FATAL and thus cannot be
caught using the PG_TRY construct. That caused the stats collector to exit.