Try to retrieve configuration variable from "post_opts".
This fixes a regression introduced by the commit 67b957d7cf. We need to ignore
the "post_opts" value when running postgres just to retrieve a GUC value
(otherwise postgres could return value of another GUC, according to
"post_opts"), but first we need to check if "post_opts" doesn't happen to
specify exactly the GUC we search for.