Do not write the encryption key to the "variables file".
In the EXEC_BACKEND branches (in practice only used on Windows) the postmaster
uses a temporary file to pass values of variables to a new child process. The
child process should read and delete the file pretty soon, but it's still a
security issue.

Now we pass the key via shared memory and only write pointer to the file. Thus
if someone steals the disk, he has no chance to get the encryption key.