Do not use PageLSN() to initialize the encryption IV.
The PageLSN() macro makes the gcc-11 compiler issue this warning:

"array subscript ‘PageHeaderData[0]’ is partly outside array bounds of
‘char[16]’"

The compiler is probably not smart enough to find out that the pd_lsn field
does fit into the tweak_loc array (or is this a compiler bug?). Nevertheless,
the use of the macro might be confusing even for human reader, so get rid of
it and use PageXLogRecPtrSet() directly.