Improve checks of the result of canAcceptConnections().
If backend starts during the CAC_OK phase, it can be one that only receives
the encryption key from the client and exits. However, so far we allowed such
a (dead_end) backend to proceed to ProcessStartupPacket() and return with
success, causing an assertion failure (due to uninitialized MyPMChildSlot).

So far the problem was only observed if the EXEC_BACKEND constat was set at
build time. The reason may be that regular fork() is fast enough to avoid this
race condition.