pg_upgrade: do not send the encryption key unnecessarily.
So far pg_upgrade tried to send the encryption key to the postmaster, in
addition to passing the -K option to pg_ctl. Since pg_ctl exited early, this
attempt only lead to an error message (failure to connect) although the
upgrade succeeded. Perhaps it was needed with older version, but not cleaned
up during the last merge.