Set the correct offset when opening a "transient" file.
The append mode is handled in a specific way for encrypted files because there
we need to rewrite the trailing part, or by using the O_APPEND flag for
unencrypted files. The code that sets the initial file offset during opening
missed the latter case.

The bug was shadowed by a bug of pwrite(), documented in "man 2 pwrite" on
both Linux and FreeBSD: this system call ignores the offset if the O_APPEND
flag was passed to open().