Fixed checking of page checksum in an encrypted cluster.
The core code uses the PageIsNew() function to determine whether page checksum
computation/verification should be skipped. However this function checks the
pd_upper field of the page header, which is encrypted, so it should not be
used on encrypted pages: the caller of PageIsNew() which does not decrypt the
page (typically the pg_checksums application) may think that the page is new
even in case the page is already initialized, but the pd_upper header field
become zero due to encryption.

To solve this, a decision was taken in the initial TDE implementation that all
pages will have the checksum in an encrypted cluster (of course, unless the
checksums are disabled globally), including the pages that only contain
zeroes. However, this decision wasn't implemented consistently in the backend,
so all-zero page (with no checksum) could appear in the database. However,
pg_checksum reported error for such pages because it expected the checksum in
*all* pages.

To make things simpler, this commit uses the assumption that a new page should
only contain zeroes. Thus we no longer need to assign checksums to the
all-zero pages: pg_checksum can distinguish the presence of a checksum by
calling PageIsNew() in an unencrypted cluster and by (a bit less cheap)
IsAllZero() in an encrypted one. The all-zero pages that already had obtained
the checksum should simply pass the checksum verification.