Fix handling of the all-visible flag in the visibility map.
So far the change of the flag did not enforce setting of a new LSN
(i.e. encryption IV) in an encrypted instance. Now that it does, it's also
necessary to write the page FPI to WAL. Otherwise the following changes would
see (LSN > RedoRecPtr) and would thus skip logging their FPI.