Do not skip WAL-logging the FPI when setting page LSN (i.e. encryption IV).
The first change of the page LSN after a checkpoint must always include
writing FPI of that page to WAL, otherwise the FPI won't be logged during the
current checkpoint cycle at all (because on the next changes the LSN will
already be greater than RedoRecPtr). The fact that encryption should not cause
a torn-page write is not important here.